+++ feature/card/src/main/java/com/itschristmas/card/cardeditor/CardEditorViewModel.kt
@@ -1,4 +1,4 @@
-package com.itschristmas.card.ui.cardeditor
+package com.itschristmas.card.cardeditor
 
 import android.util.Log
 import androidx.lifecycle.ViewModel
